backend/requirements.txt
fastapi
pydantic
uvicorn
sqlmodel
pyjwt
passlib[bcrypt]
python-multipart
python-dotenv